+++ src/TSGL/Rectangle.h
@@ -27,7 +27,7 @@ class Rectangle : public ConvexPolygon {
      *   \param outlineColor The color of the Rectangle's outline. (Defaults to BLACK)
      * \return A new Rectangle with the specified top left corner, dimensions, and color.
      */
-    Rectangle(int x, int y, int width, int height, const ColorFloat &color = BLACK, const ColorFloat& outlineColor = BLACK);
+    Rectangle(float x, float y, float width, float height, const ColorFloat &color = BLACK, const ColorFloat& outlineColor = BLACK);
 
 };
